Extract Token module
