+++ src/Data/JSON/Pointer.hs
@@ -4,28 +4,17 @@ module Data.JSON.Pointer
   , pointerFromText
   , pointerToText
   , pointerToString
-  , Token (..)
-  , tokensToText
-  , tokensToString
-  , tokensL
-  , tokenL
-  , atTokenL
   ) where
 
 import Prelude
 
-import Control.Applicative ((<|>))
-import Data.Aeson (FromJSON (..), Key, Value (..), withText)
-import Data.Aeson.Key qualified as Key
-import Data.Aeson.Optics
-import Data.Aeson.Optics.Ext
+import Data.Aeson (FromJSON (..), withText)
 import Data.Attoparsec.Text
+import Data.JSON.Pointer.Token
 import Data.List.NonEmpty (nonEmpty)
 import Data.List.NonEmpty qualified as NE
-import Data.Text (Text, pack, unpack)
+import Data.Text (Text, unpack)
 import Data.Text qualified as T
-import Optics
-import Text.Read (readEither)
 
 data Pointer
   = -- | @""@ means whole-document
@@ -50,6 +39,15 @@ pointerFromText = \case
     Nothing -> parseOnly pointerPathP p
     Just p' -> parseOnly pointerPathEndP p'
 
+pointerToText :: Pointer -> Text
+pointerToText = \case
+  PointerEmpty -> ""
+  PointerPath ts t -> tokensToText $ ts <> [t]
+  PointerPathEnd ts -> tokensToText ts <> "/-"
+
+pointerToString :: Pointer -> String
+pointerToString = unpack . pointerToText
+
 pointerPathP :: Parser Pointer
 pointerPathP = do
   ts <- maybe (fail "") pure . nonEmpty =<< pointerP
@@ -60,70 +58,3 @@ pointerPathEndP = PointerPathEnd <$> pointerP
 
 pointerP :: Parser [Token]
 pointerP = char '/' *> tokenP `sepBy1` char '/' <* endOfInput
-
-tokenP :: Parser Token
-tokenP = wtf <|> N <$> indexP <|> K <$> keyP
-
--- |
---
--- The tests love to use this value as an example. I'm clearly confused because
--- I would expect it to fail 'indexP', backtrack, then succeed in 'keyP', but it
--- causes errors with "endOfInput", so we'll special case it for now.
-wtf :: Parser Token
-wtf = K . Key.fromText <$> string "1e0"
-
-keyP :: Parser Key
-keyP =
-  Key.fromText
-    . T.replace "~0" "~"
-    . T.replace "~1" "/"
-    <$> takeTill (== '/')
-
-indexP :: Parser Int
-indexP = nonzeroP <|> 0 <$ char '0'
-
-nonzeroP :: Parser Int
-nonzeroP = do
-  ds <- (:) <$> satisfy (inClass "1-9") <*> many' digit
-  either (err ds) pure $ readEither ds
- where
-  err :: String -> String -> Parser a
-  err x msg = fail $ "Unable to read integer from " <> x <> ": " <> msg
-
-pointerToText :: Pointer -> Text
-pointerToText = \case
-  PointerEmpty -> ""
-  PointerPath ts t -> tokensToText $ ts <> [t]
-  PointerPathEnd ts -> tokensToText ts <> "/-"
-
-pointerToString :: Pointer -> String
-pointerToString = unpack . pointerToText
-
-data Token = K Key | N Int
-  deriving stock (Eq, Show)
-
-tokensToString :: [Token] -> String
-tokensToString = unpack . tokensToText
-
-tokensToText :: [Token] -> Text
-tokensToText ts = "/" <> T.intercalate "/" (map tokenToText ts <> ["-"])
-
-tokenToText :: Token -> Text
-tokenToText = \case
-  K k -> Key.toText k
-  N n -> pack $ show n
-
-tokensL :: [Token] -> AffineTraversal' Value Value
-tokensL = foldr ((%) . tokenL) $ castOptic simple
-
--- | Access a key or array index like 'ix'
-tokenL :: Token -> AffineTraversal' Value Value
-tokenL t = case t of
-  K k -> key k
-  N n -> nth n
-
--- | Access a key or array index, but 'at'-like
-atTokenL :: Token -> AffineTraversal' Value (Maybe Value)
-atTokenL = \case
-  K k -> atKey k
-  N n -> atNth n
